refactor: a custom useFrame hook
